+++ IsekaiMod/Content/ContentAdder.cs
@@ -154,6 +154,8 @@ namespace IsekaiMod.Content
                 Heritages.IsekaiAngel.AngelicBoltAbility.Add();
                 Heritages.IsekaiAngel.IsekaiAngelHeritage.Add();
                 Heritages.IsekaiVampireHeritage.Add();
+                Heritages.IsekaiSpriggan.SizeAlterationAbility.Add();
+                Heritages.IsekaiSpriggan.IsekaiSprigganHeritage.Add();
             }
             public static void AddIsekaiBackgrounds()
             {